requirements.txt
Flask==3.0.3
line-bot-sdk==3.9.0
openai==1.23.2
requests